[f] OUT-38 - Refactor upsertLatestConversation to utils
[OUT-38](https://linear.app/pdw/issue/OUT-37/implement-webhook-to-receive-label-change)
